Refactor NodeUp GossipBuilder to EtcHostsBuilder
