+++ upup/pkg/fi/nodeup/nodetasks/update_etc_hosts_task.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package dnstasks
+package nodetasks
 
 import (
 	"fmt"
